requirements.txt
applicationinsights==0.11.7
backports.csv==1.0.6
certifi==2018.11.29
chardet==3.0.4
cli-helpers==0.2.3
Click==7.0
configobj==5.0.6
Django==2.1.5
django-pyodbc-azure==2.1.0.0
djangorestframework==3.9.0
enum34==1.1.6
future==0.17.1
humanize==0.5.1
idna==2.8
mssql-cli==0.15.0
prompt-toolkit==1.0.15
Pygments==2.3.1
pyodbc==4.0.25
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
terminaltables==3.1.0
urllib3==1.24.1
wcwidth==0.1.7
